Update compiler.xml, description.html, and 10 more files... adjust Float type

+++ src/Chapter2/Num.java
@@ -1,8 +1,8 @@
 package Chapter2;
 
 public class Num extends Token {
-    public final float value;
-    public Num(float value) {
+    public final int value;
+    public Num(int value) {
         super(Tag.NUM);
         this.value = value;
     }
